dataset_merger can preserve the original dataset for further use

+++ environments/dataset_merger.py
@@ -23,6 +23,8 @@ def main():
     parser.add_argument('-f', '--force', action='store_true', default=False,
                         help='Force the merge, even if it overrides something else,' 
                              ' including the destination if it exist')
+    parser.add_argument('-rm', '--remove', action='store_true', default=False,
+                        help='Remove the original data set.')
     group = parser.add_mutually_exclusive_group()
     group.add_argument('--merge', type=str, nargs=3, metavar=('source_1', 'source_2', 'destination'),
                        default=argparse.SUPPRESS,
@@ -33,6 +35,7 @@ def main():
     if 'merge' in args:
         # let make sure everything is in order
         assert os.path.exists(args.merge[0]), "Error: dataset '{}' could not be found".format(args.merge[0])
+        assert os.path.exists(args.merge[1]), "Error: dataset '{}' could not be found".format(args.merge[1])
 
         # If the merge file exists already, delete it for the convenince of updating student's policy
         if os.path.exists(args.merge[2]) or os.path.exists(args.merge[2] + '/'):
@@ -48,13 +51,16 @@ def main():
         os.mkdir(args.merge[2])
 
         # copy files from first source
-        os.rename(args.merge[0] + "/dataset_config.json", args.merge[2] + "/dataset_config.json")
-        os.rename(args.merge[0] + "/env_globals.json", args.merge[2] + "/env_globals.json")
-
+        shutil.copy2(args.merge[0] + "/dataset_config.json", args.merge[2] + "/dataset_config.json")
+        shutil.copy2(args.merge[0] + "/env_globals.json", args.merge[2] + "/env_globals.json")
+        record = ''
         for record in sorted(glob.glob(args.merge[0] + "/record_[0-9]*/*")):
             s = args.merge[2] + "/" + record.split("/")[-2] + '/' + record.split("/")[-1]
-            os.renames(record, s)
-
+            try:
+                shutil.copy2(record, s)
+            except FileNotFoundError:  # no folders named so, we should create it first
+                os.mkdir(os.path.dirname(s))
+                shutil.copy2(record, s)
         num_episode_dataset_1 = int(record.split("/")[-2][7:]) + 1
 
         # copy files from second source
@@ -62,7 +68,11 @@ def main():
             episode = str(num_episode_dataset_1 + int(record.split("/")[-2][7:]))
             new_episode = record.split("/")[-2][:-len(episode)] + episode
             s = args.merge[2] + "/" + new_episode + '/' + record.split("/")[-1]
-            os.renames(record, s)
+            try:
+                shutil.copy2(record, s)
+            except FileNotFoundError:  # no folders named so, we should create it first
+                os.mkdir(os.path.dirname(s))
+                shutil.copy2(record, s)
         num_episode_dataset_2 = int(record.split("/")[-2][7:]) + 1
 
         # load and correct ground_truth
@@ -127,7 +137,6 @@ def main():
             for arr in prepro_load.files:
                 pr_arr = prepro_load[arr]
 
-                to_class = None
                 if arr == "episode_starts":
                     to_class = bool
                 elif arr == "actions_proba" or arr == "rewards":
@@ -151,8 +160,9 @@ def main():
         np.savez(args.merge[2] + "/preprocessed_data.npz", ** preprocessed)
 
         # remove the old folders
-        shutil.rmtree(args.merge[0])
-        shutil.rmtree(args.merge[1])
+        if args.remove:
+            shutil.rmtree(args.merge[0])
+            shutil.rmtree(args.merge[1])
 
 
 if __name__ == '__main__':
